Refactor: export weather card to a functional component
